HelmUpgrade with variables
